@@ -1,10 +1,10 @@
 name: google_ml_kit
 description: A Flutter plugin to use all APIs from Google's standalone ML Kit for mobile platforms.
-version: 0.15.0
+version: 0.16.2
 homepage: https://github.com/flutter-ml/google_ml_kit_flutter
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
+  sdk: ">=2.16.0 <4.0.0"
   flutter: ">=2.8.0"
 
 dependencies:
@@ -12,15 +12,16 @@ dependencies:
     sdk: flutter
 
   # Vision APIs
-  google_mlkit_commons: ^0.4.0
-  google_mlkit_barcode_scanning: ^0.7.0
-  google_mlkit_digital_ink_recognition: ^0.8.0
-  google_mlkit_face_detection: ^0.7.0
-  google_mlkit_image_labeling: ^0.7.0
-  google_mlkit_object_detection: ^0.8.0
-  google_mlkit_pose_detection: ^0.7.0
-  google_mlkit_selfie_segmentation: ^0.4.0
-  google_mlkit_text_recognition: ^0.7.0
+  google_mlkit_commons: ^0.5.0
+  google_mlkit_barcode_scanning: ^0.9.0
+  google_mlkit_digital_ink_recognition: ^0.9.0
+  google_mlkit_face_detection: ^0.8.0
+  google_mlkit_face_mesh_detection: ^0.0.1
+  google_mlkit_image_labeling: ^0.9.0
+  google_mlkit_object_detection: ^0.10.0
+  google_mlkit_pose_detection: ^0.9.0
+  google_mlkit_selfie_segmentation: ^0.5.0
+  google_mlkit_text_recognition: ^0.10.0
 
   # uncomment these lines when working on the local plugins
 #  google_mlkit_commons:
@@ -43,10 +44,10 @@ dependencies:
 #    path: ../google_mlkit_text_recognition
 
   # Natural Language APIs
-  google_mlkit_entity_extraction: ^0.8.0
-  google_mlkit_language_id: ^0.7.0
-  google_mlkit_translation: ^0.7.0
-  google_mlkit_smart_reply: ^0.7.0
+  google_mlkit_entity_extraction: ^0.10.0
+  google_mlkit_language_id: ^0.8.0
+  google_mlkit_translation: ^0.8.0
+  google_mlkit_smart_reply: ^0.8.0
 
   # uncomment these lines when working on the local plugins
 #  google_mlkit_entity_extraction:
@@ -61,4 +62,4 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^2.0.2
