@@ -1,14 +1,10 @@
-import 'dart:io' as io;
-
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:google_mlkit_object_detection/google_mlkit_object_detection.dart';
-import 'package:path/path.dart';
-import 'package:path_provider/path_provider.dart';
 
-import 'camera_view.dart';
+import 'detector_view.dart';
 import 'painters/object_detector_painter.dart';
+import 'utils.dart';
 
 class ObjectDetectorView extends StatefulWidget {
   @override
@@ -16,73 +12,149 @@ class ObjectDetectorView extends StatefulWidget {
 }
 
 class _ObjectDetectorView extends State<ObjectDetectorView> {
-  late ObjectDetector _objectDetector;
+  ObjectDetector? _objectDetector;
+  DetectionMode _mode = DetectionMode.stream;
   bool _canProcess = false;
   bool _isBusy = false;
   CustomPaint? _customPaint;
   String? _text;
+  var _cameraLensDirection = CameraLensDirection.back;
+  int _option = 0;
+  final _options = {
+    'default': '',
+    'object_custom': 'object_labeler.tflite',
+    'fruits': 'object_labeler_fruits.tflite',
+    'flowers': 'object_labeler_flowers.tflite',
+    'birds': 'lite-model_aiy_vision_classifier_birds_V1_3.tflite',
+    // https://tfhub.dev/google/lite-model/aiy/vision/classifier/birds_V1/3
 
-  @override
-  void initState() {
-    super.initState();
+    'food': 'lite-model_aiy_vision_classifier_food_V1_1.tflite',
+    // https://tfhub.dev/google/lite-model/aiy/vision/classifier/food_V1/1
 
-    _initializeDetector(DetectionMode.stream);
-  }
+    'plants': 'lite-model_aiy_vision_classifier_plants_V1_3.tflite',
+    // https://tfhub.dev/google/lite-model/aiy/vision/classifier/plants_V1/3
+
+    'mushrooms': 'lite-model_models_mushroom-identification_v1_1.tflite',
+    // https://tfhub.dev/bohemian-visual-recognition-alliance/lite-model/models/mushroom-identification_v1/1
+
+    'landmarks':
+        'lite-model_on_device_vision_classifier_landmarks_classifier_north_america_V1_1.tflite',
+    // https://tfhub.dev/google/lite-model/on_device_vision/classifier/landmarks_classifier_north_america_V1/1
+  };
 
   @override
   void dispose() {
     _canProcess = false;
-    _objectDetector.close();
+    _objectDetector?.close();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return CameraView(
-      title: 'Object Detector',
-      customPaint: _customPaint,
-      text: _text,
-      onImage: (inputImage) {
-        processImage(inputImage);
-      },
-      onScreenModeChanged: _onScreenModeChanged,
-      initialDirection: CameraLensDirection.back,
+    return Scaffold(
+      body: Stack(children: [
+        DetectorView(
+          title: 'Object Detector',
+          customPaint: _customPaint,
+          text: _text,
+          onImage: _processImage,
+          initialCameraLensDirection: _cameraLensDirection,
+          onCameraLensDirectionChanged: (value) => _cameraLensDirection = value,
+          onCameraFeedReady: _initializeDetector,
+          initialDetectionMode: DetectorViewMode.values[_mode.index],
+          onDetectorViewModeChanged: _onScreenModeChanged,
+        ),
+        Positioned(
+            top: 30,
+            left: 100,
+            right: 100,
+            child: Row(
+              children: [
+                Spacer(),
+                Container(
+                    decoration: BoxDecoration(
+                      color: Colors.black54,
+                      borderRadius: BorderRadius.circular(10.0),
+                    ),
+                    child: Padding(
+                      padding: const EdgeInsets.all(4.0),
+                      child: _buildDropdown(),
+                    )),
+                Spacer(),
+              ],
+            )),
+      ]),
     );
   }
 
-  void _onScreenModeChanged(ScreenMode mode) {
+  Widget _buildDropdown() => DropdownButton<int>(
+        value: _option,
+        icon: const Icon(Icons.arrow_downward),
+        elevation: 16,
+        style: const TextStyle(color: Colors.blue),
+        underline: Container(
+          height: 2,
+          color: Colors.blue,
+        ),
+        onChanged: (int? option) {
+          if (option != null) {
+            setState(() {
+              _option = option;
+              _initializeDetector();
+            });
+          }
+        },
+        items: List<int>.generate(_options.length, (i) => i)
+            .map<DropdownMenuItem<int>>((option) {
+          return DropdownMenuItem<int>(
+            value: option,
+            child: Text(_options.keys.toList()[option]),
+          );
+        }).toList(),
+      );
+
+  void _onScreenModeChanged(DetectorViewMode mode) {
     switch (mode) {
-      case ScreenMode.gallery:
-        _initializeDetector(DetectionMode.single);
+      case DetectorViewMode.gallery:
+        _mode = DetectionMode.single;
+        _initializeDetector();
         return;
 
-      case ScreenMode.liveFeed:
-        _initializeDetector(DetectionMode.stream);
+      case DetectorViewMode.liveFeed:
+        _mode = DetectionMode.stream;
+        _initializeDetector();
         return;
     }
   }
 
-  void _initializeDetector(DetectionMode mode) async {
-    print('Set detector in mode: $mode');
+  void _initializeDetector() async {
+    _objectDetector?.close();
+    _objectDetector = null;
+    print('Set detector in mode: $_mode');
 
-    // uncomment next lines if you want to use the default model
-    // final options = ObjectDetectorOptions(
-    //     mode: mode,
-    //     classifyObjects: true,
-    //     multipleObjects: true);
-    // _objectDetector = ObjectDetector(options: options);
-
-    // uncomment next lines if you want to use a local model
-    // make sure to add tflite model to assets/ml
-    final path = 'assets/ml/object_labeler.tflite';
-    final modelPath = await _getModel(path);
-    final options = LocalObjectDetectorOptions(
-      mode: mode,
-      modelPath: modelPath,
-      classifyObjects: true,
-      multipleObjects: true,
-    );
-    _objectDetector = ObjectDetector(options: options);
+    if (_option == 0) {
+      // use the default model
+      print('use the default model');
+      final options = ObjectDetectorOptions(
+        mode: _mode,
+        classifyObjects: true,
+        multipleObjects: true,
+      );
+      _objectDetector = ObjectDetector(options: options);
+    } else if (_option > 0 && _option <= _options.length) {
+      // use a custom model
+      // make sure to add tflite model to assets/ml
+      final option = _options[_options.keys.toList()[_option]] ?? '';
+      final modelPath = await getAssetPath('assets/ml/$option');
+      print('use custom model path: $modelPath');
+      final options = LocalObjectDetectorOptions(
+        mode: _mode,
+        modelPath: modelPath,
+        classifyObjects: true,
+        multipleObjects: true,
+      );
+      _objectDetector = ObjectDetector(options: options);
+    }
 
     // uncomment next lines if you want to use a remote model
     // make sure to add model to firebase
@@ -91,7 +163,7 @@ class _ObjectDetectorView extends State<ObjectDetectorView> {
     //     await FirebaseObjectDetectorModelManager().downloadModel(modelName);
     // print('Downloaded: $response');
     // final options = FirebaseObjectDetectorOptions(
-    //   mode: mode,
+    //   mode: _mode,
     //   modelName: modelName,
     //   classifyObjects: true,
     //   multipleObjects: true,
@@ -101,18 +173,24 @@ class _ObjectDetectorView extends State<ObjectDetectorView> {
     _canProcess = true;
   }
 
-  Future<void> processImage(InputImage inputImage) async {
+  Future<void> _processImage(InputImage inputImage) async {
+    if (_objectDetector == null) return;
     if (!_canProcess) return;
     if (_isBusy) return;
     _isBusy = true;
     setState(() {
       _text = '';
     });
-    final objects = await _objectDetector.processImage(inputImage);
+    final objects = await _objectDetector!.processImage(inputImage);
+    // print('Objects found: ${objects.length}\n\n');
     if (inputImage.metadata?.size != null &&
         inputImage.metadata?.rotation != null) {
       final painter = ObjectDetectorPainter(
-          objects, inputImage.metadata!.rotation, inputImage.metadata!.size);
+        objects,
+        inputImage.metadata!.size,
+        inputImage.metadata!.rotation,
+        _cameraLensDirection,
+      );
       _customPaint = CustomPaint(painter: painter);
     } else {
       String text = 'Objects found: ${objects.length}\n\n';
@@ -128,20 +206,5 @@ class _ObjectDetectorView extends State<ObjectDetectorView> {
     if (mounted) {
       setState(() {});
     }
-  }
-
-  Future<String> _getModel(String assetPath) async {
-    if (io.Platform.isAndroid) {
-      return 'flutter_assets/$assetPath';
-    }
-    final path = '${(await getApplicationSupportDirectory()).path}/$assetPath';
-    await io.Directory(dirname(path)).create(recursive: true);
-    final file = io.File(path);
-    if (!await file.exists()) {
-      final byteData = await rootBundle.load(assetPath);
-      await file.writeAsBytes(byteData.buffer
-          .asUint8List(byteData.offsetInBytes, byteData.lengthInBytes));
-    }
-    return file.path;
   }
 }
