@@ -1,4 +1,3 @@
-import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 
 import 'nlp_detector_views/entity_extraction_view.dart';
@@ -8,18 +7,15 @@ import 'nlp_detector_views/smart_reply_view.dart';
 import 'vision_detector_views/barcode_scanner_view.dart';
 import 'vision_detector_views/digital_ink_recognizer_view.dart';
 import 'vision_detector_views/face_detector_view.dart';
+import 'vision_detector_views/face_mesh_detector_view.dart';
 import 'vision_detector_views/label_detector_view.dart';
 import 'vision_detector_views/object_detector_view.dart';
 import 'vision_detector_views/pose_detector_view.dart';
 import 'vision_detector_views/selfie_segmenter_view.dart';
 import 'vision_detector_views/text_detector_view.dart';
 
-List<CameraDescription> cameras = [];
-
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
-
-  cameras = await availableCameras();
 
   runApp(MyApp());
 }
@@ -55,6 +51,7 @@ class Home extends StatelessWidget {
                     children: [
                       CustomCard('Barcode Scanning', BarcodeScannerView()),
                       CustomCard('Face Detection', FaceDetectorView()),
+                      CustomCard('Face Mesh Detection', FaceMeshDetectorView()),
                       CustomCard('Image Labeling', ImageLabelView()),
                       CustomCard('Object Detection', ObjectDetectorView()),
                       CustomCard('Text Recognition', TextRecognizerView()),
