@@ -11,9 +11,20 @@ class DigitalInkView extends StatefulWidget {
 class _DigitalInkViewState extends State<DigitalInkView> {
   final DigitalInkRecognizerModelManager _modelManager =
       DigitalInkRecognizerModelManager();
-  final String _language = 'en-US';
-  late final DigitalInkRecognizer _digitalInkRecognizer =
-      DigitalInkRecognizer(languageCode: _language);
+  var _language = 'en';
+  // Codes from https://developers.google.com/ml-kit/vision/digital-ink-recognition/base-models?hl=en#text
+  final _languages = [
+    'en',
+    'es',
+    'fr',
+    'hi',
+    'it',
+    'ja',
+    'pt',
+    'ru',
+    'zh-Hani',
+  ];
+  var _digitalInkRecognizer = DigitalInkRecognizer(languageCode: 'en');
   final Ink _ink = Ink();
   List<StrokePoint> _points = [];
   String _recognizedText = '';
@@ -31,6 +42,45 @@ class _DigitalInkViewState extends State<DigitalInkView> {
       body: SafeArea(
         child: Column(
           children: [
+            SizedBox(height: 8),
+            Padding(
+              padding: EdgeInsets.symmetric(horizontal: 16),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  _buildDropdown(),
+                  ElevatedButton(
+                    onPressed: _isModelDownloaded,
+                    child: Text('Check Model'),
+                  ),
+                  ElevatedButton(
+                    onPressed: _downloadModel,
+                    child: Icon(Icons.download),
+                  ),
+                  ElevatedButton(
+                    onPressed: _deleteModel,
+                    child: Icon(Icons.delete),
+                  ),
+                ],
+              ),
+            ),
+            SizedBox(height: 8),
+            Padding(
+              padding: EdgeInsets.symmetric(horizontal: 16),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  ElevatedButton(
+                    onPressed: _recogniseText,
+                    child: Text('Read Text'),
+                  ),
+                  ElevatedButton(
+                    onPressed: _clearPad,
+                    child: Text('Clear Pad'),
+                  ),
+                ],
+              ),
+            ),
             Expanded(
               child: GestureDetector(
                 onPanStart: (DragStartDetails details) {
@@ -69,49 +119,38 @@ class _DigitalInkViewState extends State<DigitalInkView> {
                 'Candidates: $_recognizedText',
                 style: TextStyle(fontSize: 23),
               ),
-            Padding(
-              padding: EdgeInsets.symmetric(horizontal: 16),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  ElevatedButton(
-                    onPressed: _recogniseText,
-                    child: Text('Read Text'),
-                  ),
-                  ElevatedButton(
-                    onPressed: _clearPad,
-                    child: Text('Clear Pad'),
-                  ),
-                ],
-              ),
-            ),
-            SizedBox(height: 8),
-            Padding(
-              padding: EdgeInsets.symmetric(horizontal: 16),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  ElevatedButton(
-                    onPressed: _isModelDownloaded,
-                    child: Text('Check Model'),
-                  ),
-                  ElevatedButton(
-                    onPressed: _downloadModel,
-                    child: Text('Download'),
-                  ),
-                  ElevatedButton(
-                    onPressed: _deleteModel,
-                    child: Text('Delete'),
-                  ),
-                ],
-              ),
-            ),
-            SizedBox(height: 8),
           ],
         ),
       ),
     );
   }
+
+  Widget _buildDropdown() => DropdownButton<String>(
+        value: _language,
+        icon: const Icon(Icons.arrow_downward),
+        elevation: 16,
+        style: const TextStyle(color: Colors.blue),
+        underline: Container(
+          height: 2,
+          color: Colors.blue,
+        ),
+        onChanged: (String? lang) {
+          if (lang != null) {
+            setState(() {
+              _language = lang;
+              _digitalInkRecognizer.close();
+              _digitalInkRecognizer =
+                  DigitalInkRecognizer(languageCode: _language);
+            });
+          }
+        },
+        items: _languages.map<DropdownMenuItem<String>>((lang) {
+          return DropdownMenuItem<String>(
+            value: lang,
+            child: Text(lang),
+          );
+        }).toList(),
+      );
 
   void _clearPad() {
     setState(() {
