@@ -1,13 +1,9 @@
-import 'dart:io' as io;
-
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:google_mlkit_image_labeling/google_mlkit_image_labeling.dart';
-import 'package:path/path.dart';
-import 'package:path_provider/path_provider.dart';
 
-import 'camera_view.dart';
+import 'detector_view.dart';
 import 'painters/label_detector_painter.dart';
+import 'utils.dart';
 
 class ImageLabelView extends StatefulWidget {
   @override
@@ -37,11 +33,11 @@ class _ImageLabelViewState extends State<ImageLabelView> {
 
   @override
   Widget build(BuildContext context) {
-    return CameraView(
+    return DetectorView(
       title: 'Image Labeler',
       customPaint: _customPaint,
       text: _text,
-      onImage: processImage,
+      onImage: _processImage,
     );
   }
 
@@ -52,8 +48,9 @@ class _ImageLabelViewState extends State<ImageLabelView> {
     // uncomment next lines if you want to use a local model
     // make sure to add tflite model to assets/ml
     // final path = 'assets/ml/lite-model_aiy_vision_classifier_birds_V1_3.tflite';
+    // final path = 'assets/ml/object_labeler_flowers.tflite';
     final path = 'assets/ml/object_labeler.tflite';
-    final modelPath = await _getModel(path);
+    final modelPath = await getAssetPath(path);
     final options = LocalLabelerOptions(modelPath: modelPath);
     _imageLabeler = ImageLabeler(options: options);
 
@@ -70,7 +67,7 @@ class _ImageLabelViewState extends State<ImageLabelView> {
     _canProcess = true;
   }
 
-  Future<void> processImage(InputImage inputImage) async {
+  Future<void> _processImage(InputImage inputImage) async {
     if (!_canProcess) return;
     if (_isBusy) return;
     _isBusy = true;
@@ -95,20 +92,5 @@ class _ImageLabelViewState extends State<ImageLabelView> {
     if (mounted) {
       setState(() {});
     }
-  }
-
-  Future<String> _getModel(String assetPath) async {
-    if (io.Platform.isAndroid) {
-      return 'flutter_assets/$assetPath';
-    }
-    final path = '${(await getApplicationSupportDirectory()).path}/$assetPath';
-    await io.Directory(dirname(path)).create(recursive: true);
-    final file = io.File(path);
-    if (!await file.exists()) {
-      final byteData = await rootBundle.load(assetPath);
-      await file.writeAsBytes(byteData.buffer
-          .asUint8List(byteData.offsetInBytes, byteData.lengthInBytes));
-    }
-    return file.path;
   }
 }
