@@ -1,18 +1,25 @@
-import 'dart:math';
+import 'dart:io';
 import 'dart:ui';
 import 'dart:ui' as ui;
 
+import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:google_mlkit_barcode_scanning/google_mlkit_barcode_scanning.dart';
 
 import 'coordinates_translator.dart';
 
 class BarcodeDetectorPainter extends CustomPainter {
-  BarcodeDetectorPainter(this.barcodes, this.absoluteImageSize, this.rotation);
+  BarcodeDetectorPainter(
+    this.barcodes,
+    this.imageSize,
+    this.rotation,
+    this.cameraLensDirection,
+  );
 
   final List<Barcode> barcodes;
-  final Size absoluteImageSize;
+  final Size imageSize;
   final InputImageRotation rotation;
+  final CameraLensDirection cameraLensDirection;
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -35,64 +42,83 @@ class BarcodeDetectorPainter extends CustomPainter {
       builder.addText('${barcode.displayValue}');
       builder.pop();
 
-      // Store the points for the bounding box
-      double left = double.infinity;
-      double top = double.infinity;
-      double right = double.negativeInfinity;
-      double bottom = double.negativeInfinity;
+      final left = translateX(
+        barcode.boundingBox.left,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
+      );
+      final top = translateY(
+        barcode.boundingBox.top,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
+      );
+      final right = translateX(
+        barcode.boundingBox.right,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
+      );
+      // final bottom = translateY(
+      //   barcode.boundingBox.bottom,
+      //   size,
+      //   imageSize,
+      //   rotation,
+      //   cameraLensDirection,
+      // );
+      //
+      // // Draw a bounding rectangle around the barcode
+      // canvas.drawRect(
+      //   Rect.fromLTRB(left, top, right, bottom),
+      //   paint,
+      // );
 
-      final cornerPoints = barcode.cornerPoints;
-      final boundingBox = barcode.boundingBox;
-      if (cornerPoints == null) {
-        if (boundingBox != null) {
-          left =
-              translateX(boundingBox.left, rotation, size, absoluteImageSize);
-          top = translateY(boundingBox.top, rotation, size, absoluteImageSize);
-          right =
-              translateX(boundingBox.right, rotation, size, absoluteImageSize);
-          bottom =
-              translateY(boundingBox.bottom, rotation, size, absoluteImageSize);
+      final List<Offset> cornerPoints = <Offset>[];
+      for (final point in barcode.cornerPoints) {
+        final double x = translateX(
+          point.x.toDouble(),
+          size,
+          imageSize,
+          rotation,
+          cameraLensDirection,
+        );
+        final double y = translateY(
+          point.y.toDouble(),
+          size,
+          imageSize,
+          rotation,
+          cameraLensDirection,
+        );
 
-          // Draw a bounding rectangle around the barcode
-          canvas.drawRect(
-            Rect.fromLTRB(left, top, right, bottom),
-            paint,
-          );
-        }
-      } else {
-        final List<Offset> offsetPoints = <Offset>[];
-        for (final point in cornerPoints) {
-          final double x =
-              translateX(point.x.toDouble(), rotation, size, absoluteImageSize);
-          final double y =
-              translateY(point.y.toDouble(), rotation, size, absoluteImageSize);
-
-          offsetPoints.add(Offset(x, y));
-
-          // Due to possible rotations we need to find the smallest and largest
-          top = min(top, y);
-          bottom = max(bottom, y);
-          left = min(left, x);
-          right = max(right, x);
-        }
-        // Add the first point to close the polygon
-        offsetPoints.add(offsetPoints.first);
-        canvas.drawPoints(PointMode.polygon, offsetPoints, paint);
+        cornerPoints.add(Offset(x, y));
       }
+
+      // Add the first point to close the polygon
+      cornerPoints.add(cornerPoints.first);
+      canvas.drawPoints(PointMode.polygon, cornerPoints, paint);
 
       canvas.drawParagraph(
         builder.build()
           ..layout(ParagraphConstraints(
-            width: right - left,
+            width: (right - left).abs(),
           )),
-        Offset(left, top),
+        Offset(
+            Platform.isAndroid &&
+                    cameraLensDirection == CameraLensDirection.front
+                ? right
+                : left,
+            top),
       );
     }
   }
 
   @override
   bool shouldRepaint(BarcodeDetectorPainter oldDelegate) {
-    return oldDelegate.absoluteImageSize != absoluteImageSize ||
+    return oldDelegate.imageSize != imageSize ||
         oldDelegate.barcodes != barcodes;
   }
 }
