@@ -1,36 +1,52 @@
 import 'dart:io';
 import 'dart:ui';
 
+import 'package:camera/camera.dart';
 import 'package:google_mlkit_commons/google_mlkit_commons.dart';
 
 double translateX(
-    double x, InputImageRotation rotation, Size size, Size absoluteImageSize) {
+  double x,
+  Size canvasSize,
+  Size imageSize,
+  InputImageRotation rotation,
+  CameraLensDirection cameraLensDirection,
+) {
   switch (rotation) {
     case InputImageRotation.rotation90deg:
       return x *
-          size.width /
-          (Platform.isIOS ? absoluteImageSize.width : absoluteImageSize.height);
+          canvasSize.width /
+          (Platform.isIOS ? imageSize.width : imageSize.height);
     case InputImageRotation.rotation270deg:
-      return size.width -
+      return canvasSize.width -
           x *
-              size.width /
-              (Platform.isIOS
-                  ? absoluteImageSize.width
-                  : absoluteImageSize.height);
-    default:
-      return x * size.width / absoluteImageSize.width;
+              canvasSize.width /
+              (Platform.isIOS ? imageSize.width : imageSize.height);
+    case InputImageRotation.rotation0deg:
+    case InputImageRotation.rotation180deg:
+      switch (cameraLensDirection) {
+        case CameraLensDirection.back:
+          return x * canvasSize.width / imageSize.width;
+        default:
+          return canvasSize.width - x * canvasSize.width / imageSize.width;
+      }
   }
 }
 
 double translateY(
-    double y, InputImageRotation rotation, Size size, Size absoluteImageSize) {
+  double y,
+  Size canvasSize,
+  Size imageSize,
+  InputImageRotation rotation,
+  CameraLensDirection cameraLensDirection,
+) {
   switch (rotation) {
     case InputImageRotation.rotation90deg:
     case InputImageRotation.rotation270deg:
       return y *
-          size.height /
-          (Platform.isIOS ? absoluteImageSize.height : absoluteImageSize.width);
-    default:
-      return y * size.height / absoluteImageSize.height;
+          canvasSize.height /
+          (Platform.isIOS ? imageSize.height : imageSize.width);
+    case InputImageRotation.rotation0deg:
+    case InputImageRotation.rotation180deg:
+      return y * canvasSize.height / imageSize.height;
   }
 }
