@@ -1,74 +1,134 @@
 import 'dart:math';
 
+import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:google_mlkit_face_detection/google_mlkit_face_detection.dart';
 
 import 'coordinates_translator.dart';
 
 class FaceDetectorPainter extends CustomPainter {
-  FaceDetectorPainter(this.faces, this.absoluteImageSize, this.rotation);
+  FaceDetectorPainter(
+    this.faces,
+    this.imageSize,
+    this.rotation,
+    this.cameraLensDirection,
+  );
 
   final List<Face> faces;
-  final Size absoluteImageSize;
+  final Size imageSize;
   final InputImageRotation rotation;
+  final CameraLensDirection cameraLensDirection;
 
   @override
   void paint(Canvas canvas, Size size) {
-    final Paint paint = Paint()
+    final Paint paint1 = Paint()
       ..style = PaintingStyle.stroke
       ..strokeWidth = 1.0
       ..color = Colors.red;
+    final Paint paint2 = Paint()
+      ..style = PaintingStyle.fill
+      ..strokeWidth = 1.0
+      ..color = Colors.green;
 
     for (final Face face in faces) {
+      final left = translateX(
+        face.boundingBox.left,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
+      );
+      final top = translateY(
+        face.boundingBox.top,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
+      );
+      final right = translateX(
+        face.boundingBox.right,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
+      );
+      final bottom = translateY(
+        face.boundingBox.bottom,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
+      );
+
       canvas.drawRect(
-        Rect.fromLTRB(
-          translateX(face.boundingBox.left, rotation, size, absoluteImageSize),
-          translateY(face.boundingBox.top, rotation, size, absoluteImageSize),
-          translateX(face.boundingBox.right, rotation, size, absoluteImageSize),
-          translateY(
-              face.boundingBox.bottom, rotation, size, absoluteImageSize),
-        ),
-        paint,
+        Rect.fromLTRB(left, top, right, bottom),
+        paint1,
       );
 
       void paintContour(FaceContourType type) {
-        final faceContour = face.contours[type];
-        if (faceContour?.points != null) {
-          for (final Point point in faceContour!.points) {
+        final contour = face.contours[type];
+        if (contour?.points != null) {
+          for (final Point point in contour!.points) {
             canvas.drawCircle(
                 Offset(
                   translateX(
-                      point.x.toDouble(), rotation, size, absoluteImageSize),
+                    point.x.toDouble(),
+                    size,
+                    imageSize,
+                    rotation,
+                    cameraLensDirection,
+                  ),
                   translateY(
-                      point.y.toDouble(), rotation, size, absoluteImageSize),
+                    point.y.toDouble(),
+                    size,
+                    imageSize,
+                    rotation,
+                    cameraLensDirection,
+                  ),
                 ),
                 1,
-                paint);
+                paint1);
           }
         }
       }
 
-      paintContour(FaceContourType.face);
-      paintContour(FaceContourType.leftEyebrowTop);
-      paintContour(FaceContourType.leftEyebrowBottom);
-      paintContour(FaceContourType.rightEyebrowTop);
-      paintContour(FaceContourType.rightEyebrowBottom);
-      paintContour(FaceContourType.leftEye);
-      paintContour(FaceContourType.rightEye);
-      paintContour(FaceContourType.upperLipTop);
-      paintContour(FaceContourType.upperLipBottom);
-      paintContour(FaceContourType.lowerLipTop);
-      paintContour(FaceContourType.lowerLipBottom);
-      paintContour(FaceContourType.noseBridge);
-      paintContour(FaceContourType.noseBottom);
-      paintContour(FaceContourType.leftCheek);
-      paintContour(FaceContourType.rightCheek);
+      void paintLandmark(FaceLandmarkType type) {
+        final landmark = face.landmarks[type];
+        if (landmark?.position != null) {
+          canvas.drawCircle(
+              Offset(
+                translateX(
+                  landmark!.position.x.toDouble(),
+                  size,
+                  imageSize,
+                  rotation,
+                  cameraLensDirection,
+                ),
+                translateY(
+                  landmark.position.y.toDouble(),
+                  size,
+                  imageSize,
+                  rotation,
+                  cameraLensDirection,
+                ),
+              ),
+              2,
+              paint2);
+        }
+      }
+
+      for (final type in FaceContourType.values) {
+        paintContour(type);
+      }
+
+      for (final type in FaceLandmarkType.values) {
+        paintLandmark(type);
+      }
     }
   }
 
   @override
   bool shouldRepaint(FaceDetectorPainter oldDelegate) {
-    return oldDelegate.absoluteImageSize != absoluteImageSize ||
-        oldDelegate.faces != faces;
+    return oldDelegate.imageSize != imageSize || oldDelegate.faces != faces;
   }
 }
