@@ -1,6 +1,8 @@
+import 'dart:io';
 import 'dart:ui';
 import 'dart:ui' as ui;
 
+import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:google_mlkit_text_recognition/google_mlkit_text_recognition.dart';
 
@@ -8,11 +10,16 @@ import 'coordinates_translator.dart';
 
 class TextRecognizerPainter extends CustomPainter {
   TextRecognizerPainter(
-      this.recognizedText, this.absoluteImageSize, this.rotation);
+    this.recognizedText,
+    this.imageSize,
+    this.rotation,
+    this.cameraLensDirection,
+  );
 
   final RecognizedText recognizedText;
-  final Size absoluteImageSize;
+  final Size imageSize;
   final InputImageRotation rotation;
+  final CameraLensDirection cameraLensDirection;
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -36,25 +43,137 @@ class TextRecognizerPainter extends CustomPainter {
       builder.pop();
 
       final left = translateX(
-          textBlock.boundingBox.left, rotation, size, absoluteImageSize);
-      final top = translateY(
-          textBlock.boundingBox.top, rotation, size, absoluteImageSize);
-      final right = translateX(
-          textBlock.boundingBox.right, rotation, size, absoluteImageSize);
-      final bottom = translateY(
-          textBlock.boundingBox.bottom, rotation, size, absoluteImageSize);
-
-      canvas.drawRect(
-        Rect.fromLTRB(left, top, right, bottom),
-        paint,
+        textBlock.boundingBox.left,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
       );
+      final top = translateY(
+        textBlock.boundingBox.top,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
+      );
+      final right = translateX(
+        textBlock.boundingBox.right,
+        size,
+        imageSize,
+        rotation,
+        cameraLensDirection,
+      );
+      // final bottom = translateY(
+      //   textBlock.boundingBox.bottom,
+      //   size,
+      //   imageSize,
+      //   rotation,
+      //   cameraLensDirection,
+      // );
+      //
+      // canvas.drawRect(
+      //   Rect.fromLTRB(left, top, right, bottom),
+      //   paint,
+      // );
+
+      final List<Offset> cornerPoints = <Offset>[];
+      for (final point in textBlock.cornerPoints) {
+        double x = translateX(
+          point.x.toDouble(),
+          size,
+          imageSize,
+          rotation,
+          cameraLensDirection,
+        );
+        double y = translateY(
+          point.y.toDouble(),
+          size,
+          imageSize,
+          rotation,
+          cameraLensDirection,
+        );
+
+        if (Platform.isAndroid) {
+          switch (cameraLensDirection) {
+            case CameraLensDirection.front:
+              switch (rotation) {
+                case InputImageRotation.rotation0deg:
+                case InputImageRotation.rotation90deg:
+                  break;
+                case InputImageRotation.rotation180deg:
+                  x = size.width - x;
+                  y = size.height - y;
+                  break;
+                case InputImageRotation.rotation270deg:
+                  x = translateX(
+                    point.y.toDouble(),
+                    size,
+                    imageSize,
+                    rotation,
+                    cameraLensDirection,
+                  );
+                  y = size.height -
+                      translateY(
+                        point.x.toDouble(),
+                        size,
+                        imageSize,
+                        rotation,
+                        cameraLensDirection,
+                      );
+                  break;
+              }
+              break;
+            case CameraLensDirection.back:
+              switch (rotation) {
+                case InputImageRotation.rotation0deg:
+                case InputImageRotation.rotation270deg:
+                  break;
+                case InputImageRotation.rotation180deg:
+                  x = size.width - x;
+                  y = size.height - y;
+                  break;
+                case InputImageRotation.rotation90deg:
+                  x = size.width -
+                      translateX(
+                        point.y.toDouble(),
+                        size,
+                        imageSize,
+                        rotation,
+                        cameraLensDirection,
+                      );
+                  y = translateY(
+                    point.x.toDouble(),
+                    size,
+                    imageSize,
+                    rotation,
+                    cameraLensDirection,
+                  );
+                  break;
+              }
+              break;
+            case CameraLensDirection.external:
+              break;
+          }
+        }
+
+        cornerPoints.add(Offset(x, y));
+      }
+
+      // Add the first point to close the polygon
+      cornerPoints.add(cornerPoints.first);
+      canvas.drawPoints(PointMode.polygon, cornerPoints, paint);
 
       canvas.drawParagraph(
         builder.build()
           ..layout(ParagraphConstraints(
-            width: right - left,
+            width: (right - left).abs(),
           )),
-        Offset(left, top),
+        Offset(
+            Platform.isAndroid &&
+                    cameraLensDirection == CameraLensDirection.front
+                ? right
+                : left,
+            top),
       );
     }
   }
