@@ -1,7 +1,8 @@
+import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:google_mlkit_text_recognition/google_mlkit_text_recognition.dart';
 
-import 'camera_view.dart';
+import 'detector_view.dart';
 import 'painters/text_detector_painter.dart';
 
 class TextRecognizerView extends StatefulWidget {
@@ -10,12 +11,13 @@ class TextRecognizerView extends StatefulWidget {
 }
 
 class _TextRecognizerViewState extends State<TextRecognizerView> {
-  final TextRecognizer _textRecognizer =
-      TextRecognizer(script: TextRecognitionScript.latin);
+  var _script = TextRecognitionScript.latin;
+  var _textRecognizer = TextRecognizer(script: TextRecognitionScript.latin);
   bool _canProcess = true;
   bool _isBusy = false;
   CustomPaint? _customPaint;
   String? _text;
+  var _cameraLensDirection = CameraLensDirection.back;
 
   @override
   void dispose() async {
@@ -26,17 +28,67 @@ class _TextRecognizerViewState extends State<TextRecognizerView> {
 
   @override
   Widget build(BuildContext context) {
-    return CameraView(
-      title: 'Text Detector',
-      customPaint: _customPaint,
-      text: _text,
-      onImage: (inputImage) {
-        processImage(inputImage);
-      },
+    return Scaffold(
+      body: Stack(children: [
+        DetectorView(
+          title: 'Text Detector',
+          customPaint: _customPaint,
+          text: _text,
+          onImage: _processImage,
+          initialCameraLensDirection: _cameraLensDirection,
+          onCameraLensDirectionChanged: (value) => _cameraLensDirection = value,
+        ),
+        Positioned(
+            top: 30,
+            left: 100,
+            right: 100,
+            child: Row(
+              children: [
+                Spacer(),
+                Container(
+                    decoration: BoxDecoration(
+                      color: Colors.black54,
+                      borderRadius: BorderRadius.circular(10.0),
+                    ),
+                    child: Padding(
+                      padding: const EdgeInsets.all(4.0),
+                      child: _buildDropdown(),
+                    )),
+                Spacer(),
+              ],
+            )),
+      ]),
     );
   }
 
-  Future<void> processImage(InputImage inputImage) async {
+  Widget _buildDropdown() => DropdownButton<TextRecognitionScript>(
+        value: _script,
+        icon: const Icon(Icons.arrow_downward),
+        elevation: 16,
+        style: const TextStyle(color: Colors.blue),
+        underline: Container(
+          height: 2,
+          color: Colors.blue,
+        ),
+        onChanged: (TextRecognitionScript? script) {
+          if (script != null) {
+            setState(() {
+              _script = script;
+              _textRecognizer.close();
+              _textRecognizer = TextRecognizer(script: _script);
+            });
+          }
+        },
+        items: TextRecognitionScript.values
+            .map<DropdownMenuItem<TextRecognitionScript>>((script) {
+          return DropdownMenuItem<TextRecognitionScript>(
+            value: script,
+            child: Text(script.name),
+          );
+        }).toList(),
+      );
+
+  Future<void> _processImage(InputImage inputImage) async {
     if (!_canProcess) return;
     if (_isBusy) return;
     _isBusy = true;
@@ -46,8 +98,12 @@ class _TextRecognizerViewState extends State<TextRecognizerView> {
     final recognizedText = await _textRecognizer.processImage(inputImage);
     if (inputImage.metadata?.size != null &&
         inputImage.metadata?.rotation != null) {
-      final painter = TextRecognizerPainter(recognizedText,
-          inputImage.metadata!.size, inputImage.metadata!.rotation);
+      final painter = TextRecognizerPainter(
+        recognizedText,
+        inputImage.metadata!.size,
+        inputImage.metadata!.rotation,
+        _cameraLensDirection,
+      );
       _customPaint = CustomPaint(painter: painter);
     } else {
       _text = 'Recognized text:\n\n${recognizedText.text}';
