@@ -155,7 +155,8 @@
                     } else if ([entity.entityType isEqualToString: MLKEntityExtractionEntityTypeDateTime]) {
                         type = 2;
                         entityData[@"dateTimeGranularity"] = @(entity.dateTimeEntity.dateTimeGranularity);
-                        entityData[@"timestamp"] = @(entity.dateTimeEntity.dateTime.timeIntervalSince1970);
+                        // result is expected in milliseconds, not seconds.
+                        entityData[@"timestamp"] = @(entity.dateTimeEntity.dateTime.timeIntervalSince1970 * 1000);
                     } else if ([entity.entityType isEqualToString: MLKEntityExtractionEntityTypeEmail]) {
                         type = 3;
                     } else if ([entity.entityType isEqualToString: MLKEntityExtractionEntityTypeFlightNumber]) {
