@@ -137,7 +137,7 @@ public class ObjectDetector implements MethodChannel.MethodCallHandler {
         String path = (String) options.get("path");
 
         LocalModel localModel = new LocalModel.Builder()
-                .setAssetFilePath(path)
+                .setAbsoluteFilePath(path)
                 .build();
 
         CustomObjectDetectorOptions.Builder builder = new CustomObjectDetectorOptions.Builder(localModel);
